azure_container_app/using_managed_identity/requirements.txt
fastapi
uvicorn[standard]
azure-identity
azure-storage-blob
pymongo[srv]
